requirements.txt
pytest
playwright
boto3
nova-act
pytest-html
python-dotenv
